chore: remove stalebot
Stalebots are user host hostile, requiring the reporters to
babysit their issues and eventually result in issues being closed
even though the problem is still present in the codebase.

It also sends a "go away" message to people trying to contribute
reporting issues.

An open issue for a problem is way better than *still having that problem*
but with no way to track it.